Proper handling of invalid DTO received, Single DTO for page request that also has filter, fixed AdvertisementRepository page retrieving query as it wasn't working for null filter items,  conversion of Spring Page to CustomPage, added archivation mechanism to project, SaltedPassword is not used anymore, as Spring PasswordEncoder doesn't need it. Username and Password validation. Password now has list of allow characters.

+++ source/backend/src/main/kotlin/cz/opendatalab/egidio/backend/shared/pagination/CustomPage.kt
@@ -19,7 +19,7 @@ data class CustomPage<T>(
      * Total number of items available for retrieval (pages count = ceil(totalItemsAvailable / size))
      */
     @PositiveOrZero
-    val totalItemsAvailable: Int,
+    val totalItemsAvailable: Long,
 
     val items: List<T>
 )
